Removed baseTrigger
